Refactor tvConnectionManager methods to use generic TV object methods
